+++ plugin_tooling/src/com/github/rustdt/tooling/RustBuildOutputParser.java
@@ -10,17 +10,16 @@
  *******************************************************************************/
 package com.github.rustdt.tooling;
 
-import java.io.BufferedReader;
-import java.io.IOException;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
 import melnorme.lang.tooling.ops.BuildOutputParser;
+import melnorme.lang.utils.parse.StringParseSource;
 
 
 public abstract class RustBuildOutputParser extends BuildOutputParser {
 	
-	protected static final Pattern ERROR_LINE_Regex = Pattern.compile(
+	protected static final Pattern MESSAGE_LINE_Regex = Pattern.compile(
 		"^([^:\\n]*):" + // file
 		"(\\d*):((\\d*):)?" +// line:column
 		"( (\\d*):(\\d*))?" + // end line:column
@@ -30,12 +29,12 @@ public abstract class RustBuildOutputParser extends BuildOutputParser {
 	
 	
 	@Override
-	protected void doParseLine(String outputLine, BufferedReader br) throws IOException {
+	protected void doParseLine(String outputLine, StringParseSource output) {
 		if(!outputLine.contains(":")) {
 			return; // Ignore line
 		}
 		
-		Matcher matcher = ERROR_LINE_Regex.matcher(outputLine);
+		Matcher matcher = MESSAGE_LINE_Regex.matcher(outputLine);
 		if(!matcher.matches()) {
 			handleUnknownLineSyntax(outputLine);
 			return;
@@ -51,6 +50,41 @@ public abstract class RustBuildOutputParser extends BuildOutputParser {
 		String messageTypeString = matcher.group(8);
 		
 		String message = matcher.group(9);
+		
+		
+		while(true) {
+			String nextLine = output.stringUntil("\n");
+			
+			if(nextLine.isEmpty() || MESSAGE_LINE_Regex.matcher(nextLine).matches()) {
+				break;
+			}
+			
+			if(nextLine.startsWith("error: aborting due to previous error")) {
+				// We reached the end of messages, exhaust remaining tool output.
+				while(output.consume() != -1) {
+				}
+				break;
+			}
+			
+			// We assume this is a multi line message.
+			output.consumeAhead(nextLine);
+			output.tryConsume("\n"); /* FIXME: consume Line */
+			
+			// However, first try to determine if this is the source range component, 
+			// which we don't need
+			
+			String thirdLine = output.stringUntil("\n");
+			String thirdLineTrimmed = thirdLine.trim();
+			if(thirdLineTrimmed.startsWith("^") && 
+					(thirdLineTrimmed.endsWith("^") || thirdLineTrimmed.endsWith("~"))) {
+				output.consumeAhead(thirdLine);
+				output.tryConsume("\n"); /* FIXME: consume Line */
+				// dont add this message, or nextLine, to actual error message.
+				break;
+			}
+			
+			message += "\n" + nextLine;
+		}
 		
 		addMessage(pathString, lineString, columnString, endLineString, endColumnString, messageTypeString, message);
 	}