+++ documentation/ChangeLog.md
@@ -9,6 +9,7 @@
  * Fixed #31: Added workaround to fix OS X issue "Could not execute process `rustc -vV`". 
  * Fixed #27: the preference pages are now searchable in the Preferences dialog search field, by means of relevant keywords.
  * Fixed: `Tab policy: "Spaces Only"` preference ignored when pressing TAB.
+ * Fixed #39: Unable to parse multiline rustc error messages
  
 
 ### 0.2.0